moved example flowsheets
